+++ composeApp/src/commonMain/kotlin/storage/KeyValueStorage.kt
@@ -1,22 +1,23 @@
 package storage
 
+import kotlinx.serialization.Serializable
+
 interface KeyValueStorage {
     var isOnboardingCompleted: Boolean
     var isSignInCompleted: Boolean
-//    var isSignedOut: Boolean
-//    var signInInfo: LoginInfo?
+    var signInInfo: LoginInfo?
 //    val observableToken: Flow<String>
 
     fun cleanStorage()
 }
 
-//@Serializable
-//data class LoginInfo(val username: String, val password: Int)
+@Serializable
+data class LoginInfo(val username: String, val password: String)
 
 enum class StorageKeys {
     ONBOARDING_INFO,
-    SIGNIN_INFO;
-//    SIGNOUT_INFO;
+    SIGNIN_INFO,
+    LOGIN_INFO;
 
     val key get() = this.name
 }
