+++ composeApp/src/commonMain/kotlin/ui/CommonBackground.kt
@@ -1,0 +1,48 @@
+package ui
+
+import AppColors
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import kotlinproject.composeapp.generated.resources.Res
+import kotlinproject.composeapp.generated.resources.background_main
+import kotlinproject.composeapp.generated.resources.manrope_bold
+import org.jetbrains.compose.resources.Font
+import org.jetbrains.compose.resources.StringResource
+import org.jetbrains.compose.resources.painterResource
+import org.jetbrains.compose.resources.stringResource
+
+@Composable
+fun CommonBackground(text: StringResource) {
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+    ) {
+        Image(
+            painter = painterResource(Res.drawable.background_main),
+            contentDescription = null,
+            modifier = Modifier
+                .fillMaxSize(),
+            contentScale = ContentScale.FillBounds
+        )
+
+        Text(
+            modifier = Modifier
+                .align(Alignment.TopStart)
+                .padding(top = 50.dp, start = 16.dp),
+            text = stringResource(text),
+            color = AppColors.White,
+            fontFamily = FontFamily(Font(Res.font.manrope_bold)),
+            fontSize = 32.sp,
+        )
+    }
+}
