--- composeApp/src/commonMain/kotlin/ui/ElementsSizing.kt
@@ -1,15 +1,0 @@
-package ui
-
-const val designHeight = 812F // Screen height (Figma)
-const val designWidth = 375F // Screen width (Figma)
-
-object ElementsSizing {
-     val headerHeightMultiplier = 32 / designHeight
-     val headerWidthMultiplier = 68 / designWidth
-
-     val executionerSizeMultiplier = 220 / designHeight
-     val tabHeightMultiplier = 68 / designHeight
-
-    val signOutHeightMultiplier = 48 / designHeight
-     val signOutWidthMultiplier = 343 / designWidth
-}
