+++ gradle/libs.versions.toml
@@ -1,6 +1,6 @@
 [versions]
 accompanistPager = "0.36.0"
-agp = "8.8.0" #DO NOT CHANGE YET! THERE IS NO SUPPORT of v8.8.0 on MAC
+agp = "8.7.3" #DO NOT CHANGE YET! THERE IS NO SUPPORT of v8.8.0 on MAC
 android-compileSdk = "34"
 android-minSdk = "24"
 android-targetSdk = "34"
@@ -8,7 +8,7 @@ androidx-activityCompose = "1.10.0"
 androidxMaterial = "1.7.6"
 androidxUiAndroid = "1.8.0-alpha08"
 annotationJvm = "1.9.1"
-compose-plugin = "1.6.11"
+compose-plugin = "1.6.10"
 core = "1.15.0"
 coroutinesVersion = "1.8.0"
 foundationLayoutAndroid = "1.7.6"
@@ -16,7 +16,7 @@ koin = "4.0.0"
 kotlin = "2.0.20"
 lifecycleLivedataCoreKtx = "2.8.7"
 multiplatformSettings = "1.0.0"
-serialization = "1.4.1"
+serialization = "1.6.0"
 ui = "1.7.6"
 uiAndroid = "1.7.6"
 uiGraphicsAndroid = "1.7.6"
